testing

+++ src/main/java/com/adm/journalApp/controllerV2/service/JournalEntryService.java
@@ -5,6 +5,8 @@ import java.util.List;
 import java.util.Optional;
 
 import org.bson.types.ObjectId;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
@@ -16,6 +18,8 @@ import com.adm.journalApp.repository.JournalEntryRepository;
 
 @Service
 public class JournalEntryService {
+
+    private static final Logger logger = LoggerFactory.getLogger(JournalEntryService.class);
 
     @Autowired
 
@@ -32,10 +36,10 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
 
         } catch (Exception e) {
-            System.out.println(e);
+            logger.info("hahahahahha");
             throw new RuntimeException("An error occured while saving the entry.", e);
         }
     }
@@ -53,10 +57,24 @@ public class JournalEntryService {
     }
 
     @Transactional
-    public void deleteById(ObjectId id, String userName) {
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        if(removed) {
+            userService.saveUser(user);
+            journalEntryRepository.deleteById(id);
+        }
+        }catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting the entry.", e);
+        }
+        return removed;
+    }
+
+    public List<JournalEntry> findByUserName(String userName) {
         User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+        return journalEntryRepository.findByUserId(user.getId());
     }
 }
